+++ packages/common/src/types/field.ts
@@ -261,6 +261,7 @@ export const isCustomSqlDimension = (
 export type CompiledCustomSqlDimension = CustomSqlDimension & {
     compiledSql: string;
     tablesReferences: Array<string>;
+    parameterReferences?: string[];
 };
 
 export type CompiledCustomDimension =
@@ -377,6 +378,7 @@ export interface Field {
     urls?: FieldUrl[];
     index?: number;
     tags?: string[];
+    parameterReferences?: string[];
 }
 
 export const isField = (field: AnyType): field is Field =>
@@ -445,14 +447,16 @@ export interface Dimension extends Field {
     aiHint?: string;
 }
 
-export interface CompiledDimension extends Dimension {
+type CompiledProperties = {
     compiledSql: string; // sql string with resolved template variables
     tablesReferences: Array<string> | undefined;
     tablesRequiredAttributes?: Record<
         string,
         Record<string, string | string[]>
     >;
-}
+};
+export type CompiledDimension = Dimension & CompiledProperties;
+export type CompiledMetric = Metric & CompiledProperties;
 
 export type CompiledField = CompiledDimension | CompiledMetric;
 
@@ -460,15 +464,6 @@ export const isDimension = (
     field: ItemsMap[string] | AdditionalMetric | undefined, // NOTE: `ItemsMap converts AdditionalMetric to Metric
 ): field is Dimension =>
     isField(field) && field.fieldType === FieldType.DIMENSION;
-
-export interface CompiledMetric extends Metric {
-    compiledSql: string;
-    tablesReferences: Array<string> | undefined;
-    tablesRequiredAttributes?: Record<
-        string,
-        Record<string, string | string[]>
-    >;
-}
 
 export interface FilterableDimension extends Dimension {
     type: