+++ packages/frontend/src/components/Explorer/index.tsx
@@ -1,6 +1,7 @@
 import { Stack } from '@mantine/core';
 import { memo, type FC } from 'react';
 import { useParams } from 'react-router';
+import { useCompiledSql } from '../../hooks/useCompiledSql';
 import { useExplore } from '../../hooks/useExplore';
 import useExplorerContext from '../../providers/Explorer/useExplorerContext';
 import { DrillDownModal } from '../MetricQueryData/DrillDownModal';
@@ -11,6 +12,7 @@ import { CustomMetricModal } from './CustomMetricModal';
 import ExplorerHeader from './ExplorerHeader';
 import FiltersCard from './FiltersCard/FiltersCard';
 import { FormatModal } from './FormatModal';
+import ParametersCard from './ParametersCard/ParametersCard';
 import ResultsCard from './ResultsCard/ResultsCard';
 import SqlCard from './SqlCard/SqlCard';
 import VisualizationCard from './VisualizationCard/VisualizationCard';
@@ -35,6 +37,10 @@ const Explorer: FC<{ hideHeader?: boolean }> = memo(
 
         const { data: explore } = useExplore(unsavedChartVersionTableName);
 
+        const { data: { parameterReferences } = {} } = useCompiledSql({
+            enabled: !!unsavedChartVersionTableName,
+        });
+
         return (
             <MetricQueryDataProvider
                 metricQuery={unsavedChartVersionMetricQuery}
@@ -44,6 +50,14 @@ const Explorer: FC<{ hideHeader?: boolean }> = memo(
             >
                 <Stack sx={{ flexGrow: 1 }}>
                     {!hideHeader && isEditMode && <ExplorerHeader />}
+
+                    {!!unsavedChartVersionTableName &&
+                        parameterReferences &&
+                        parameterReferences?.length > 0 && (
+                            <ParametersCard
+                                parameterReferences={parameterReferences}
+                            />
+                        )}
 
                     <FiltersCard />